Use nodeunit for spellcheck
