+++ .travis.yml
@@ -1,7 +1,3 @@
-sudo: false
 language: node_js
 node_js:
-  - "0.11"
-
-script:
-  - npm test
+- 0.10
