+++ gulpfile.js
@@ -1,0 +1,4 @@
+var gulp = require('gulp');
+var buildTools = require('node-firefox-build-tools');
+
+buildTools.loadGulpTasks(gulp);
